Rework install script internals
The script now breaks down into two functions, archTest and nodeSelect, which collectively test for the presence of a working NodeJS installation of sufficiently recent vintage or the ability to use the included node executable. The end result is that the script now executes without needing to change directory and the actual call to node that acts on `app/install.js` uses variables defined by the functions to call on the right interpreter with the necessary arguments without case logic and the script will error out with actionable messages before reaching it if installation is impossible.

The test for $OSTYPE = darwin is removed since the release of Apple Silicon means that systems with ARM architecture can now meet the test but can't use the included x64 node.